Moved user testing utility methods to a separate utility class
