+++ app/src/androidTest/java/ch/epfl/sdp/peakar/general/MainMenuActivityTest.java
@@ -37,7 +37,7 @@ import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
-import static ch.epfl.sdp.peakar.user.AuthAccountTest.removeAuthUser;
+import static ch.epfl.sdp.peakar.utils.UserTestHelper.removeAuthUser;
 
 @RunWith(AndroidJUnit4.class)
 public class MainMenuActivityTest {
